+++ add.php
@@ -1,5 +1,12 @@
 <?php
-require_once 'bootstrap.php';
+/**
+ * @var mysqli $connection идентификатор соединения
+ * @var string $title заголовок страницы
+ * @var string $user_name имя пользователя
+ * @var int $is_auth флаг авторизации
+ */
+
+require_once __DIR__ . '/bootstrap.php';
 
 $categories = get_categories($connection);
 $lots = get_active_lots($connection);
@@ -8,8 +15,8 @@ $errors = [];
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
-    $form_data = filter_form_fields($_POST);
-    $errors = validate_lot_form($form_data);
+    $form_data = filter_form_lot($_POST);
+    $errors = validate_lot_form($form_data, $_FILES);
 
     if (empty($errors)) {
         $form_data['image_url'] = upload_file($_FILES);
@@ -18,10 +25,11 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     }
 }
 
+$main_menu = include_template('/menu/top_menu.php', ['categories' => $categories]);
 $main_page = include_template('add.php', ['error' => $errors, 'categories' => $categories]);
 $main_footer = include_template('footer.php', ['categories' => $categories]);
-$layout_content = include_template('layout2.php', [
-    'categories' => $categories,
+$layout_content = include_template('layout.php', [
+    'top_menu' => $main_menu,
     'content' => $main_page,
     'footer' => $main_footer,
     'title' => $title,
