+++ css/style.css
@@ -1253,3 +1253,7 @@ input:invalid .form__error{
 .rates__item--end {
   opacity: 0.4;
 }
+
+.page__item--hidden {
+  display: none;
+}
