+++ templates/footer.php
@@ -1,9 +1,14 @@
+<?php
+/**
+ * @var array $categories массив с категориями
+ */
+?>
 <nav class="nav">
         <ul class="nav__list container">
             <!--заполните этот список из массива категорий-->
             <?php foreach($categories as $category): ?>
             <li class="nav__item">
-                <a href="pages/all-lots.html"><?= $category['name']?></a>
+                <a href="/pages/all-lots.html"><?= $category['name']?></a>
             </li>
             <?php endforeach; ?>
         </ul>
