+++ templates/menu/top_menu.php
@@ -1,0 +1,9 @@
+<nav class="nav">
+    <ul class="nav__list container">
+        <?php foreach ($categories as $category) : ?>
+            <li class="nav__item">
+                <a href="all-lots.html"><?= $category['name'] ?></a>
+            </li>
+        <?php endforeach; ?>
+    </ul>
+</nav>
